@@ -1,51 +1,85 @@
-// Autogenerated from Pigeon (v0.1.9), do not edit directly.
+// Autogenerated from Pigeon (v1.0.1), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
-// ignore_for_file: public_member_api_docs, non_constant_identifier_names, avoid_as, unused_import
-// @dart = 2.8
+// ignore_for_file: public_member_api_docs, non_constant_identifier_names, avoid_as, unused_import, unnecessary_parenthesis, prefer_null_aware_operators, omit_local_variable_types, unused_shown_name
+// @dart = 2.12
 import 'dart:async';
-import 'package:flutter/services.dart';
 import 'dart:typed_data' show Uint8List, Int32List, Int64List, Float64List;
 
+import 'package:flutter/foundation.dart' show WriteBuffer, ReadBuffer;
+import 'package:flutter/services.dart';
+
 class Version {
-  String string;
-  // ignore: unused_element
-  Map<dynamic, dynamic> _toMap() {
-    final Map<dynamic, dynamic> pigeonMap = <dynamic, dynamic>{};
+  String? string;
+
+  Object encode() {
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
     pigeonMap['string'] = string;
     return pigeonMap;
   }
-  // ignore: unused_element
-  static Version _fromMap(Map<dynamic, dynamic> pigeonMap) {
-    if (pigeonMap == null){
-      return null;
+
+  static Version decode(Object message) {
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
+    return Version()..string = pigeonMap['string'] as String?;
+  }
+}
+
+class _ExampleApiCodec extends StandardMessageCodec {
+  const _ExampleApiCodec();
+  @override
+  void writeValue(WriteBuffer buffer, Object? value) {
+    if (value is Version) {
+      buffer.putUint8(128);
+      writeValue(buffer, value.encode());
+    } else {
+      super.writeValue(buffer, value);
     }
-    final Version result = Version();
-    result.string = pigeonMap['string'];
-    return result;
+  }
+
+  @override
+  Object? readValueOfType(int type, ReadBuffer buffer) {
+    switch (type) {
+      case 128:
+        return Version.decode(readValue(buffer)!);
+
+      default:
+        return super.readValueOfType(type, buffer);
+    }
   }
 }
 
 class ExampleApi {
+  /// Constructor for [ExampleApi].  The [binaryMessenger] named argument is
+  /// available for dependency injection.  If it is left null, the default
+  /// BinaryMessenger will be used which routes to the host platform.
+  ExampleApi({BinaryMessenger? binaryMessenger})
+      : _binaryMessenger = binaryMessenger;
+
+  final BinaryMessenger? _binaryMessenger;
+
+  static const MessageCodec<Object?> codec = _ExampleApiCodec();
+
   Future<Version> getPlatformVersion() async {
-    const BasicMessageChannel<dynamic> channel =
-        BasicMessageChannel<dynamic>('dev.flutter.pigeon.ExampleApi.getPlatformVersion', StandardMessageCodec());
-    
-    final Map<dynamic, dynamic> replyMap = await channel.send(null);
+    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+        'dev.flutter.pigeon.ExampleApi.getPlatformVersion', codec,
+        binaryMessenger: _binaryMessenger);
+    final Map<Object?, Object?>? replyMap =
+        await channel.send(null) as Map<Object?, Object?>?;
     if (replyMap == null) {
       throw PlatformException(
         code: 'channel-error',
         message: 'Unable to establish connection on channel.',
-        details: null);
+        details: null,
+      );
     } else if (replyMap['error'] != null) {
-      final Map<dynamic, dynamic> error = replyMap['error'];
+      final Map<Object?, Object?> error =
+          (replyMap['error'] as Map<Object?, Object?>?)!;
       throw PlatformException(
-          code: error['code'],
-          message: error['message'],
-          details: error['details']);
+        code: (error['code'] as String?)!,
+        message: error['message'] as String?,
+        details: error['details'],
+      );
     } else {
-      return Version._fromMap(replyMap['result']);
+      return (replyMap['result'] as Version?)!;
     }
-    
   }
 }
-
