@@ -1,16 +1,23 @@
-// Autogenerated from Pigeon (v0.1.9), do not edit directly.
+// Autogenerated from Pigeon (v1.0.1), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 
 package com.example.pigeon_plugin;
 
 import io.flutter.plugin.common.BasicMessageChannel;
 import io.flutter.plugin.common.BinaryMessenger;
+import io.flutter.plugin.common.MessageCodec;
 import io.flutter.plugin.common.StandardMessageCodec;
+import java.io.ByteArrayOutputStream;
+import java.nio.ByteBuffer;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 /** Generated class from Pigeon. */
-@SuppressWarnings("unused")
+@SuppressWarnings(
+    {"unused", "unchecked", "CodeBlock2Expr", "RedundantSuppression"})
 public class Messages {
 
   /** Generated class from Pigeon that represents data sent in messages. */
@@ -19,36 +26,68 @@ public class Messages {
     public String getString() { return string; }
     public void setString(String setterArg) { this.string = setterArg; }
 
-    HashMap toMap() {
-      HashMap<String, Object> toMapResult = new HashMap<>();
+    Map<String, Object> toMap() {
+      Map<String, Object> toMapResult = new HashMap<>();
       toMapResult.put("string", string);
       return toMapResult;
     }
-    static Version fromMap(HashMap map) {
+    static Version fromMap(Map<String, Object> map) {
       Version fromMapResult = new Version();
       Object string = map.get("string");
       fromMapResult.string = (String)string;
       return fromMapResult;
     }
   }
+  private static class ExampleApiCodec extends StandardMessageCodec {
+    public static final ExampleApiCodec INSTANCE = new ExampleApiCodec();
+    private ExampleApiCodec() {}
+    @Override
+    protected Object readValueOfType(byte type, ByteBuffer buffer) {
+      switch (type) {
+      case (byte)128:
+        return Version.fromMap((Map<String, Object>)readValue(buffer));
 
-  /** Generated interface from Pigeon that represents a handler of messages from Flutter.*/
+      default:
+        return super.readValueOfType(type, buffer);
+      }
+    }
+    @Override
+    protected void writeValue(ByteArrayOutputStream stream, Object value) {
+      if (value instanceof Version) {
+        stream.write(128);
+        writeValue(stream, ((Version)value).toMap());
+      } else {
+        super.writeValue(stream, value);
+      }
+    }
+  }
+
+  /**
+   * Generated interface from Pigeon that represents a handler of messages from
+   * Flutter.
+   */
   public interface ExampleApi {
     Version getPlatformVersion();
 
-    /** Sets up an instance of `ExampleApi` to handle messages through the `binaryMessenger` */
+    /** The codec used by ExampleApi. */
+    static MessageCodec<Object> getCodec() { return ExampleApiCodec.INSTANCE; }
+
+    /**
+     * Sets up an instance of `ExampleApi` to handle messages through the
+     * `binaryMessenger`.
+     */
     static void setup(BinaryMessenger binaryMessenger, ExampleApi api) {
       {
-        BasicMessageChannel<Object> channel =
-            new BasicMessageChannel<>(binaryMessenger, "dev.flutter.pigeon.ExampleApi.getPlatformVersion", new StandardMessageCodec());
+        BasicMessageChannel<Object> channel = new BasicMessageChannel<>(
+            binaryMessenger, "dev.flutter.pigeon.ExampleApi.getPlatformVersion",
+            getCodec());
         if (api != null) {
           channel.setMessageHandler((message, reply) -> {
-            HashMap<String, HashMap> wrapped = new HashMap<>();
+            Map<String, Object> wrapped = new HashMap<>();
             try {
               Version output = api.getPlatformVersion();
-              wrapped.put("result", output.toMap());
-            }
-            catch (Exception exception) {
+              wrapped.put("result", output);
+            } catch (Error | RuntimeException exception) {
               wrapped.put("error", wrapError(exception));
             }
             reply.reply(wrapped);
@@ -59,10 +98,10 @@ public class Messages {
       }
     }
   }
-  private static HashMap wrapError(Exception exception) {
-    HashMap<String, Object> errorMap = new HashMap<>();
+  private static Map<String, Object> wrapError(Throwable exception) {
+    Map<String, Object> errorMap = new HashMap<>();
     errorMap.put("message", exception.toString());
-    errorMap.put("code", null);
+    errorMap.put("code", exception.getClass().getSimpleName());
     errorMap.put("details", null);
     return errorMap;
   }
