@@ -1,7 +1,8 @@
-// Autogenerated from Pigeon (v0.1.9), do not edit directly.
+// Autogenerated from Pigeon (v1.0.1), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 #import <Foundation/Foundation.h>
 @protocol FlutterBinaryMessenger;
+@protocol FlutterMessageCodec;
 @class FlutterError;
 @class FlutterStandardTypedData;
 
@@ -13,10 +14,13 @@ NS_ASSUME_NONNULL_BEGIN
 @property(nonatomic, copy, nullable) NSString * string;
 @end
 
+/// The codec used by FLTExampleApi.
+NSObject<FlutterMessageCodec> *FLTExampleApiGetCodec(void);
+
 @protocol FLTExampleApi
--(nullable FLTVersion *)getPlatformVersion:(FlutterError *_Nullable *_Nonnull)error;
+- (nullable FLTVersion *)getPlatformVersionWithError:(FlutterError *_Nullable *_Nonnull)error;
 @end
 
-extern void FLTExampleApiSetup(id<FlutterBinaryMessenger> binaryMessenger, id<FLTExampleApi> _Nullable api);
+extern void FLTExampleApiSetup(id<FlutterBinaryMessenger> binaryMessenger, NSObject<FLTExampleApi> *_Nullable api);
 
 NS_ASSUME_NONNULL_END
