@@ -1,4 +1,4 @@
-// Autogenerated from Pigeon (v0.1.9), do not edit directly.
+// Autogenerated from Pigeon (v1.0.1), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 #import "messages.h"
 #import <Flutter/Flutter.h>
@@ -7,54 +7,113 @@
 #error File requires ARC to be enabled.
 #endif
 
-static NSDictionary* wrapResult(NSDictionary *result, FlutterError *error) {
+static NSDictionary<NSString *, id> *wrapResult(id result,
+                                                FlutterError *error) {
   NSDictionary *errorDict = (NSDictionary *)[NSNull null];
   if (error) {
-    errorDict = [NSDictionary dictionaryWithObjectsAndKeys:
-        (error.code ? error.code : [NSNull null]), @"code",
-        (error.message ? error.message : [NSNull null]), @"message",
-        (error.details ? error.details : [NSNull null]), @"details",
-        nil];
+    errorDict = @{
+      @"code" : (error.code ? error.code : [NSNull null]),
+      @"message" : (error.message ? error.message : [NSNull null]),
+      @"details" : (error.details ? error.details : [NSNull null]),
+    };
   }
-  return [NSDictionary dictionaryWithObjectsAndKeys:
-      (result ? result : [NSNull null]), @"result",
-      errorDict, @"error",
-      nil];
+  return @{
+    @"result" : (result ? result : [NSNull null]),
+    @"error" : errorDict,
+  };
 }
 
 @interface FLTVersion ()
-+(FLTVersion*)fromMap:(NSDictionary*)dict;
--(NSDictionary*)toMap;
++ (FLTVersion *)fromMap:(NSDictionary *)dict;
+- (NSDictionary *)toMap;
 @end
 
 @implementation FLTVersion
-+(FLTVersion*)fromMap:(NSDictionary*)dict {
-  FLTVersion* result = [[FLTVersion alloc] init];
++ (FLTVersion *)fromMap:(NSDictionary *)dict {
+  FLTVersion *result = [[FLTVersion alloc] init];
   result.string = dict[@"string"];
   if ((NSNull *)result.string == [NSNull null]) {
     result.string = nil;
   }
   return result;
 }
--(NSDictionary*)toMap {
-  return [NSDictionary dictionaryWithObjectsAndKeys:(self.string ? self.string : [NSNull null]), @"string", nil];
+- (NSDictionary *)toMap {
+  return [NSDictionary
+      dictionaryWithObjectsAndKeys:(self.string ? self.string : [NSNull null]),
+                                   @"string", nil];
 }
 @end
 
-void FLTExampleApiSetup(id<FlutterBinaryMessenger> binaryMessenger, id<FLTExampleApi> api) {
+@interface FLTExampleApiCodecReader : FlutterStandardReader
+@end
+@implementation FLTExampleApiCodecReader
+- (nullable id)readValueOfType:(UInt8)type {
+  switch (type) {
+  case 128:
+    return [FLTVersion fromMap:[self readValue]];
+
+  default:
+    return [super readValueOfType:type];
+  }
+}
+@end
+
+@interface FLTExampleApiCodecWriter : FlutterStandardWriter
+@end
+@implementation FLTExampleApiCodecWriter
+- (void)writeValue:(id)value {
+  if ([value isKindOfClass:[FLTVersion class]]) {
+    [self writeByte:128];
+    [self writeValue:[value toMap]];
+  } else {
+    [super writeValue:value];
+  }
+}
+@end
+
+@interface FLTExampleApiCodecReaderWriter : FlutterStandardReaderWriter
+@end
+@implementation FLTExampleApiCodecReaderWriter
+- (FlutterStandardWriter *)writerWithData:(NSMutableData *)data {
+  return [[FLTExampleApiCodecWriter alloc] initWithData:data];
+}
+- (FlutterStandardReader *)readerWithData:(NSData *)data {
+  return [[FLTExampleApiCodecReader alloc] initWithData:data];
+}
+@end
+
+NSObject<FlutterMessageCodec> *FLTExampleApiGetCodec() {
+  static dispatch_once_t s_pred = 0;
+  static FlutterStandardMessageCodec *s_sharedObject = nil;
+  dispatch_once(&s_pred, ^{
+    FLTExampleApiCodecReaderWriter *readerWriter =
+        [[FLTExampleApiCodecReaderWriter alloc] init];
+    s_sharedObject =
+        [FlutterStandardMessageCodec codecWithReaderWriter:readerWriter];
+  });
+  return s_sharedObject;
+}
+
+void FLTExampleApiSetup(id<FlutterBinaryMessenger> binaryMessenger,
+                        NSObject<FLTExampleApi> *api) {
   {
-    FlutterBasicMessageChannel *channel =
-      [FlutterBasicMessageChannel
-        messageChannelWithName:@"dev.flutter.pigeon.ExampleApi.getPlatformVersion"
-        binaryMessenger:binaryMessenger];
+    FlutterBasicMessageChannel *channel = [FlutterBasicMessageChannel
+        messageChannelWithName:
+            @"dev.flutter.pigeon.ExampleApi.getPlatformVersion"
+               binaryMessenger:binaryMessenger
+                         codec:FLTExampleApiGetCodec()];
     if (api) {
-      [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
-        FlutterError *error;
-        FLTVersion *output = [api getPlatformVersion:&error];
-        callback(wrapResult([output toMap], error));
-      }];
-    }
-    else {
+      NSCAssert(
+          [api respondsToSelector:@selector(getPlatformVersionWithError:)],
+          @"FLTExampleApi api doesn't respond to "
+          @"@selector(getPlatformVersionWithError:)");
+      [channel
+          setMessageHandler:^(id _Nullable message, FlutterReply callback) {
+            FlutterError *error;
+            FLTVersion *output = [api getPlatformVersionWithError:&error];
+            callback(wrapResult(output, error));
+          }];
+    } else {
       [channel setMessageHandler:nil];
     }
   }
